Updates graphql-go to lab259
Because of the subscriptions implementation, some changes had to be made
in the graphql-go/graphql implementation. These changes are simple, but,
were not merged into the main repository (and not sure if it will).

Hence, I had to fork this repository just to use the lab259's graphql
fork.

After the development is done, probably this repository "will be no
more", as soon as the lab259's fork get merged (if it will).